Initial changes regarding fix for anchor links when converting to HTML.
Need to review failed unit tests and add new ones if needed.

We should compare this with the new repo that is currently active:

https://github.com/EricWhiteDev/Open-Xml-PowerTools

Also, we should review whether or not we should be using the vNext branch or keep it with the Master branch in the new EricWhiteDev repo.